Move validation engine to Medidata.Cloud.ExcelLoader assembly
