+++ Medidata.Cloud.ExcelLoader.Tests/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="AutoFixture" version="3.37.3" targetFramework="net4" />
+  <package id="AutoFixture.AutoRhinoMocks" version="3.37.3" targetFramework="net4" />
+  <package id="RhinoMocks" version="3.6.1" targetFramework="net4" />
+</packages>
