+++ Medidata.Cloud.ExcelLoader/Medidata.Cloud.ExcelLoader.csproj
@@ -39,6 +39,10 @@
       <HintPath>..\packages\ImpromptuInterface.6.2.2\lib\net40\ImpromptuInterface.dll</HintPath>
       <Private>True</Private>
     </Reference>
+    <Reference Include="Ploeh.AutoFixture, Version=3.37.3.0, Culture=neutral, PublicKeyToken=b24654c590009d4f, processorArchitecture=MSIL">
+      <HintPath>..\packages\AutoFixture.3.37.3\lib\net40\Ploeh.AutoFixture.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="Microsoft.CSharp" />
@@ -57,6 +61,7 @@
     <Compile Include="CellTypeConverters\DoubleConverter.cs" />
     <Compile Include="CellTypeConverters\FloatConverter.cs" />
     <Compile Include="CellTypeConverters\ICellTypeValueConverter.cs" />
+    <Compile Include="Helpers\StringValidationMessageExtensions.cs" />
     <Compile Include="IAutoFilterProvider.cs" />
     <Compile Include="ICellTypeValueConverterFactory.cs" />
     <Compile Include="CellTypeConverters\IntConverter.cs" />
@@ -78,6 +83,19 @@
     <Compile Include="ExcelParser.cs" />
     <Compile Include="SheetBuilder.cs" />
     <Compile Include="SheetParser.cs" />
+    <Compile Include="Validations\DefaultSequentialRuleValidator.cs" />
+    <Compile Include="Validations\IValidationError.cs" />
+    <Compile Include="Validations\IValidationMessage.cs" />
+    <Compile Include="Validations\IValidationResult.cs" />
+    <Compile Include="Validations\IValidationRule.cs" />
+    <Compile Include="Validations\IValidationRuleResult.cs" />
+    <Compile Include="Validations\IValidationWarning.cs" />
+    <Compile Include="Validations\IValidator.cs" />
+    <Compile Include="Validations\Rules\ValidationRuleBase.cs" />
+    <Compile Include="Validations\Rules\ValidationRuleResult.cs" />
+    <Compile Include="Validations\ValidationError.cs" />
+    <Compile Include="Validations\ValidationResult.cs" />
+    <Compile Include="Validations\ValidationWarning.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Medidata.Cloud.ExcelLoader.nuspec" />
