+++ Medidata.Cloud.ExcelLoader/Validations/IValidationError.cs
@@ -1,0 +1,6 @@
+namespace Medidata.Cloud.ExcelLoader.Validations
+{
+    public interface IValidationError : IValidationMessage
+    {
+    }
+}
