+++ Medidata.Cloud.ExcelLoader/Validations/IValidationMessage.cs
@@ -1,0 +1,9 @@
+namespace Medidata.Cloud.ExcelLoader.Validations
+{
+    public interface IValidationMessage
+    {
+        IValidationRule ByWhom { get; }
+        string Message { get; }
+        string Where { get; }
+    }
+}
