+++ Medidata.Cloud.ExcelLoader/Validations/IValidationResult.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace Medidata.Cloud.ExcelLoader.Validations
+{
+    public interface IValidationResult
+    {
+        IExcelParser ValidationTarget { get; }
+        IList<IValidationMessage> Messages { get; }
+    }
+}
