+++ Medidata.Cloud.ExcelLoader/Validations/IValidationRule.cs
@@ -1,0 +1,7 @@
+namespace Medidata.Cloud.ExcelLoader.Validations
+{
+    public interface IValidationRule
+    {
+        IValidationRuleResult Check(IExcelParser blockPlan);
+    }
+}
