+++ Medidata.Cloud.ExcelLoader/Validations/IValidationRuleResult.cs
@@ -1,0 +1,8 @@
+namespace Medidata.Cloud.ExcelLoader.Validations
+{
+    public interface IValidationRuleResult
+    {
+        IValidationMessage Message { get; }
+        bool ShouldContinue { get; }
+    }
+}
