+++ Medidata.Cloud.ExcelLoader/Validations/IValidationWarning.cs
@@ -1,0 +1,6 @@
+namespace Medidata.Cloud.ExcelLoader.Validations
+{
+    public interface IValidationWarning : IValidationMessage
+    {
+    }
+}
