+++ Medidata.Cloud.ExcelLoader/Validations/IValidator.cs
@@ -1,0 +1,7 @@
+namespace Medidata.Cloud.ExcelLoader.Validations
+{
+    public interface IValidator
+    {
+        IValidationResult Validate(IExcelParser excelParser);
+    }
+}
