+++ Medidata.Cloud.ExcelLoader/Validations/Rules/ValidationRuleResult.cs
@@ -1,0 +1,8 @@
+namespace Medidata.Cloud.ExcelLoader.Validations.Rules
+{
+    internal class ValidationRuleResult : IValidationRuleResult
+    {
+        public IValidationMessage Message { get; set; }
+        public bool ShouldContinue { get; set; }
+    }
+}
