+++ Medidata.Cloud.ExcelLoader/Validations/ValidationResult.cs
@@ -1,0 +1,15 @@
+using System.Collections.Generic;
+
+namespace Medidata.Cloud.ExcelLoader.Validations
+{
+    internal class ValidationResult : IValidationResult
+    {
+        public ValidationResult()
+        {
+            Messages = new List<IValidationMessage>();
+        }
+
+        public IExcelParser ValidationTarget { get; set; }
+        public IList<IValidationMessage> Messages { get; private set; }
+    }
+}
