+++ Medidata.Cloud.ExcelLoader/packages.config
@@ -1,8 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-
 <packages>
-  <package id="CreateNewNuGetPackageFromProjectAfterEachBuild" version="1.8.8" targetFramework="net40"
-           developmentDependency="true" />
+  <package id="AutoFixture" version="3.37.3" targetFramework="net4" />
+  <package id="CreateNewNuGetPackageFromProjectAfterEachBuild" version="1.8.8" targetFramework="net40" developmentDependency="true" />
   <package id="DocumentFormat.OpenXml" version="2.5" targetFramework="net40" />
   <package id="ImpromptuInterface" version="6.2.2" targetFramework="net40" />
 </packages>
