+++ Medidata.Rave.Tsdv.Loader.Tests/packages.config
@@ -1,10 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-
 <packages>
-  <package id="AutoFixture" version="3.36.10" targetFramework="net45" />
-  <package id="AutoFixture.AutoRhinoMocks" version="3.36.10" targetFramework="net45" />
+  <package id="AutoFixture" version="3.37.3" targetFramework="net4" />
+  <package id="AutoFixture.AutoRhinoMocks" version="3.37.3" targetFramework="net4" />
   <package id="DocumentFormat.OpenXml" version="2.5" targetFramework="net45" />
   <package id="Medidata.Interfaces" version="0.1.3" targetFramework="net40" />
-  <package id="RhinoMocks" version="3.6" targetFramework="net45" />
+  <package id="RhinoMocks" version="3.6.1" targetFramework="net4" />
   <package id="YAXLib" version="2.13" targetFramework="net45" />
 </packages>
