+++ Medidata.Rave.Tsdv.Loader/Medidata.Rave.Tsdv.Loader.csproj
@@ -59,7 +59,6 @@
     <Compile Include="SheetDefinitions\v1\TsdvReportLoader.cs" />
     <Compile Include="SheetDefinitions\TsdvReportGenericBuilder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Helpers\StringValidationMessageExtensions.cs" />
     <Compile Include="Resource.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -73,20 +72,7 @@
     <Compile Include="SheetDefinitions\v1\ITierField.cs" />
     <Compile Include="SheetDefinitions\v1\ITierFolder.cs" />
     <Compile Include="SheetDefinitions\v1\ITierForm.cs" />
-    <Compile Include="Validation\DefaultSequentialRuleValidator.cs" />
-    <Compile Include="Validation\IValidationError.cs" />
-    <Compile Include="Validation\IValidationMessage.cs" />
-    <Compile Include="Validation\IValidationResult.cs" />
-    <Compile Include="Validation\IValidationRule.cs" />
-    <Compile Include="Validation\IValidationRuleResult.cs" />
-    <Compile Include="Validation\IValidationWarning.cs" />
-    <Compile Include="Validation\IValidator.cs" />
-    <Compile Include="Validation\Rules\LocalizableValidationRuleBase.cs" />
-    <Compile Include="Validation\Rules\ValidationRuleBase.cs" />
-    <Compile Include="Validation\Rules\ValidationRuleResult.cs" />
-    <Compile Include="Validation\ValidationError.cs" />
-    <Compile Include="Validation\ValidationResult.cs" />
-    <Compile Include="Validation\ValidationWarning.cs" />
+    <Compile Include="Validations\Rules\LocalizableValidationRuleBase.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Medidata.Rave.Tsdv.Loader.nuspec" />
