--- Medidata.Rave.Tsdv.Loader/Validation/IValidationError.cs
@@ -1,6 +1,0 @@
-namespace Medidata.Rave.Tsdv.Loader.Validation
-{
-    public interface IValidationError : IValidationMessage
-    {
-    }
-}
