--- Medidata.Rave.Tsdv.Loader/Validation/IValidationMessage.cs
@@ -1,9 +1,0 @@
-namespace Medidata.Rave.Tsdv.Loader.Validation
-{
-    public interface IValidationMessage
-    {
-        IValidationRule ByWhom { get; }
-        string Message { get; }
-        string Where { get; }
-    }
-}
