--- Medidata.Rave.Tsdv.Loader/Validation/IValidationRule.cs
@@ -1,9 +1,0 @@
-using Medidata.Interfaces.TSDV;
-
-namespace Medidata.Rave.Tsdv.Loader.Validation
-{
-    public interface IValidationRule
-    {
-        IValidationRuleResult Check(IBlockPlan blockPlan);
-    }
-}
