--- Medidata.Rave.Tsdv.Loader/Validation/IValidationRuleResult.cs
@@ -1,8 +1,0 @@
-namespace Medidata.Rave.Tsdv.Loader.Validation
-{
-    public interface IValidationRuleResult
-    {
-        IValidationMessage Message { get; }
-        bool ShouldContinue { get; }
-    }
-}
