--- Medidata.Rave.Tsdv.Loader/Validation/IValidationWarning.cs
@@ -1,6 +1,0 @@
-namespace Medidata.Rave.Tsdv.Loader.Validation
-{
-    public interface IValidationWarning : IValidationMessage
-    {
-    }
-}
