--- Medidata.Rave.Tsdv.Loader/Validation/IValidator.cs
@@ -1,9 +1,0 @@
-using Medidata.Interfaces.TSDV;
-
-namespace Medidata.Rave.Tsdv.Loader.Validation
-{
-    public interface IValidator
-    {
-        IValidationResult Validate(IBlockPlan blockPlan);
-    }
-}
