--- Medidata.Rave.Tsdv.Loader/Validation/Rules/ValidationRuleResult.cs
@@ -1,8 +1,0 @@
-namespace Medidata.Rave.Tsdv.Loader.Validation.Rules
-{
-    internal class ValidationRuleResult : IValidationRuleResult
-    {
-        public IValidationMessage Message { get; set; }
-        public bool ShouldContinue { get; set; }
-    }
-}
